Rework pagination, fix caption2 font, some logging fixes, updated packages

+++ utils/pagination/pagination.js
@@ -1,6 +1,6 @@
 import MessageCollector from "./awaitmessages.js";
 import InteractionCollector from "./awaitinteractions.js";
-import fetch from "node-fetch";
+import { ComponentInteraction } from "eris";
 
 export default async (client, message, pages, timeout = 120000) => {
   const manageMessages = message.channel.guild && message.channel.permissionsOf(client.user.id).has("manageMessages") ? true : false;
@@ -63,36 +63,47 @@ export default async (client, message, pages, timeout = 120000) => {
       ]
     }]
   };
-  const ackOptions = {
-    method: "POST",
-    body: "{\"type\":6}",
-    headers: {
-      "Content-Type": "application/json"
-    }
-  };
   let currentPage = await client.createMessage(message.channel.id, Object.assign(pages[page], options, pages.length > 1 ? components : {}));
   if (pages.length > 1) {
-    const interactionCollector = new InteractionCollector(client, currentPage, timeout);
-    interactionCollector.on("interaction", async (interaction, id, token, member) => {
-      if (member === message.author.id) {
-        switch (interaction) {
+    const interactionCollector = new InteractionCollector(client, currentPage, ComponentInteraction, timeout);
+    interactionCollector.on("interaction", async (interaction) => {
+      if ((interaction.member ?? interaction.user).id === message.author.id) {
+        switch (interaction.data.custom_id) {
           case "back":
-            await fetch(`https://discord.com/api/v9/interactions/${id}/${token}/callback`, ackOptions);
+            await interaction.deferUpdate();
             page = page > 0 ? --page : pages.length - 1;
             currentPage = await currentPage.edit(Object.assign(pages[page], options));
             break;
           case "forward":
-            await fetch(`https://discord.com/api/v9/interactions/${id}/${token}/callback`, ackOptions);
+            await interaction.deferUpdate();
             page = page + 1 < pages.length ? ++page : 0;
             currentPage = await currentPage.edit(Object.assign(pages[page], options));
             break;
           case "jump":
-            await fetch(`https://discord.com/api/v9/interactions/${id}/${token}/callback`, ackOptions);
+            await interaction.deferUpdate();
             const newComponents = JSON.parse(JSON.stringify(components));
             for (const index of newComponents.components[0].components.keys()) {
               newComponents.components[0].components[index].disabled = true;
             }
             currentPage = await currentPage.edit(newComponents);
+            const jumpComponents = {
+              components: [{
+                type: 1,
+                components: [{
+                  type: 3,
+                  custom_id: "seekDropdown",
+                  placeholder: "Page Number",
+                  options: []
+                }]
+              }]
+            };
+            for (let i = 0; i < pages.length; i++) {
+              const payload = {
+                label: i + 1,
+                value: i
+              };
+              jumpComponents.components[0].components[0].options[i] = payload;
+            }
             client.createMessage(message.channel.id, Object.assign({ content: "What page do you want to jump to?" }, {
               messageReference: {
                 channelID: currentPage.channel.id,
@@ -103,15 +114,20 @@ export default async (client, message, pages, timeout = 120000) => {
               allowedMentions: {
                 repliedUser: false
               }
-            })).then(askMessage => {
-              const messageCollector = new MessageCollector(client, askMessage.channel, (response) => response.author.id === message.author.id && !isNaN(response.content) && Number(response.content) <= pages.length && Number(response.content) > 0, {
-                time: timeout,
-                maxMatches: 1
-              });
-              return messageCollector.on("message", async (response) => {
+            }, jumpComponents)).then(askMessage => {
+              const dropdownCollector = new InteractionCollector(client, askMessage, ComponentInteraction, timeout);
+              let ended = false;
+              dropdownCollector.on("interaction", async (response) => {
+                if (response.data.custom_id !== "seekDropdown") return;
                 if (await client.getMessage(askMessage.channel.id, askMessage.id).catch(() => undefined)) await askMessage.delete();
-                if (manageMessages) await response.delete();
-                page = Number(response.content) - 1;
+                page = Number(response.data.values[0]);
+                currentPage = await currentPage.edit(Object.assign(pages[page], options, components));
+                ended = true;
+                dropdownCollector.stop();
+              });
+              dropdownCollector.once("end", async () => {
+                if (ended) return;
+                if (await client.getMessage(askMessage.channel.id, askMessage.id).catch(() => undefined)) await askMessage.delete();
                 currentPage = await currentPage.edit(Object.assign(pages[page], options, components));
               });
             }).catch(error => {
@@ -119,7 +135,7 @@ export default async (client, message, pages, timeout = 120000) => {
             });
             break;
           case "delete":
-            await fetch(`https://discord.com/api/v9/interactions/${id}/${token}/callback`, ackOptions);
+            await interaction.deferUpdate();
             interactionCollector.emit("end");
             if (await client.getMessage(currentPage.channel.id, currentPage.id).catch(() => undefined)) await currentPage.delete();
             return;
@@ -130,10 +146,10 @@ export default async (client, message, pages, timeout = 120000) => {
     });
     interactionCollector.once("end", async () => {
       interactionCollector.removeAllListeners("interaction");
+      for (const index of components.components[0].components.keys()) {
+        components.components[0].components[index].disabled = true;
+      }
       if (await client.getMessage(currentPage.channel.id, currentPage.id).catch(() => undefined)) {
-        for (const index of components.components[0].components.keys()) {
-          components.components[0].components[index].disabled = true;
-        }
         await currentPage.edit(components);
       }
     });
